+++ flutter/lib/src/native/platform/constraints.g.dart
@@ -15,7 +15,8 @@ void Function(Object, int, ForeignValue)? _setVideoConstraintValue;
 void Function(Object, int, ForeignValue)? _setAudioConstraintValue;
 void Function(Object, int, Object)? _setVideoConstraint;
 void Function(Object, int, Object)? _setDisplayVideoConstraint;
-void Function(Object, int, Object)? _setAudioConstraint;
+void Function(Object, int, Object)? _setDeviceAudioConstraint;
+void Function(Object, int, Object)? _setDisplayAudioConstraint;
 
 _ErrorSetterFnDart? _constraints__init_device_constraints__set_error;
 _ErrorSetterFnDart? _constraints__init_display_constraints__set_error;
@@ -25,7 +26,8 @@ _ErrorSetterFnDart? _constraints__set_video_constraint_value__set_error;
 _ErrorSetterFnDart? _constraints__set_audio_constraint_value__set_error;
 _ErrorSetterFnDart? _constraints__set_video_constraint__set_error;
 _ErrorSetterFnDart? _constraints__set_display_video_constraint__set_error;
-_ErrorSetterFnDart? _constraints__set_audio_constraint__set_error;
+_ErrorSetterFnDart? _constraints__set_device_audio_constraint__set_error;
+_ErrorSetterFnDart? _constraints__set_display_audio_constraint__set_error;
 
 void registerFunction(
   DynamicLibrary dl, {
@@ -37,7 +39,8 @@ void registerFunction(
   required void Function(Object, int, ForeignValue) setAudioConstraintValue,
   required void Function(Object, int, Object) setVideoConstraint,
   required void Function(Object, int, Object) setDisplayVideoConstraint,
-  required void Function(Object, int, Object) setAudioConstraint,
+  required void Function(Object, int, Object) setDeviceAudioConstraint,
+  required void Function(Object, int, Object) setDisplayAudioConstraint,
 }) {
   _initDeviceConstraints = initDeviceConstraints;
   _initDisplayConstraints = initDisplayConstraints;
@@ -47,7 +50,8 @@ void registerFunction(
   _setAudioConstraintValue = setAudioConstraintValue;
   _setVideoConstraint = setVideoConstraint;
   _setDisplayVideoConstraint = setDisplayVideoConstraint;
-  _setAudioConstraint = setAudioConstraint;
+  _setDeviceAudioConstraint = setDeviceAudioConstraint;
+  _setDisplayAudioConstraint = setDisplayAudioConstraint;
 
   _constraints__init_device_constraints__set_error = dl
       .lookupFunction<_ErrorSetterFnC, _ErrorSetterFnDart>(
@@ -81,9 +85,13 @@ void registerFunction(
       .lookupFunction<_ErrorSetterFnC, _ErrorSetterFnDart>(
         'constraints__set_display_video_constraint__set_error',
       );
-  _constraints__set_audio_constraint__set_error = dl
+  _constraints__set_device_audio_constraint__set_error = dl
       .lookupFunction<_ErrorSetterFnC, _ErrorSetterFnDart>(
-        'constraints__set_audio_constraint__set_error',
+        'constraints__set_device_audio_constraint__set_error',
+      );
+  _constraints__set_display_audio_constraint__set_error = dl
+      .lookupFunction<_ErrorSetterFnC, _ErrorSetterFnDart>(
+        'constraints__set_display_audio_constraint__set_error',
       );
 
   Pointer<NativeFunction<Handle Function()>> initDeviceConstraints_native =
@@ -109,10 +117,17 @@ void registerFunction(
     _setDisplayVideoConstraintProxy,
   );
   Pointer<NativeFunction<Void Function(Handle, Int64, Handle)>>
-  setAudioConstraint_native = Pointer.fromFunction(_setAudioConstraintProxy);
+  setDeviceAudioConstraint_native = Pointer.fromFunction(
+    _setDeviceAudioConstraintProxy,
+  );
+  Pointer<NativeFunction<Void Function(Handle, Int64, Handle)>>
+  setDisplayAudioConstraint_native = Pointer.fromFunction(
+    _setDisplayAudioConstraintProxy,
+  );
 
   dl.lookupFunction<
     Void Function(
+      Pointer,
       Pointer,
       Pointer,
       Pointer,
@@ -133,6 +148,7 @@ void registerFunction(
       Pointer,
       Pointer,
       Pointer,
+      Pointer,
     )
   >('register_constraints')(
     initDeviceConstraints_native,
@@ -143,7 +159,8 @@ void registerFunction(
     setAudioConstraintValue_native,
     setVideoConstraint_native,
     setDisplayVideoConstraint_native,
-    setAudioConstraint_native,
+    setDeviceAudioConstraint_native,
+    setDisplayAudioConstraint_native,
   );
 }
 
@@ -219,11 +236,20 @@ void _setDisplayVideoConstraintProxy(Object arg0, int arg1, Object arg2) {
   }
 }
 
-void _setAudioConstraintProxy(Object arg0, int arg1, Object arg2) {
+void _setDeviceAudioConstraintProxy(Object arg0, int arg1, Object arg2) {
   try {
-    return _setAudioConstraint!(arg0, arg1, arg2);
+    return _setDeviceAudioConstraint!(arg0, arg1, arg2);
   } catch (e) {
-    _constraints__set_audio_constraint__set_error!(e);
+    _constraints__set_device_audio_constraint__set_error!(e);
+    return;
+  }
+}
+
+void _setDisplayAudioConstraintProxy(Object arg0, int arg1, Object arg2) {
+  try {
+    return _setDisplayAudioConstraint!(arg0, arg1, arg2);
+  } catch (e) {
+    _constraints__set_display_audio_constraint__set_error!(e);
     return;
   }
 }