+++ flutter/CHANGELOG.md
@@ -14,10 +14,19 @@ All user visible changes to this project will be documented in this file. This p
 
 - Bumped up minimum supported [Flutter] version to 3.35. ([#231])
 - Bumped up minimum supported [Dart] version to 3.9. ([#231])
+- Rename `MediaStreamSettings.audio()` to `MediaStreamSettings.deviceAudio()`. ([#225])
+- `RoomHandle.muteAudio()`, `RoomHandle.unmuteAudio()`, `RoomHandle.enableAudio()`, `RoomHandle.disableAudio()`, `RoomHandle.enableRemoteAudio()`, and `RoomHandle.disableRemoteAudio()` now receive an argument to specify `MediaSourceKind`. ([#225])
 
 ### Added
 
+- `MediaStreamSettings.displayAudio()` to configure system audio capture. ([#225])
 - `ConnectionHandle.getState()` and `ConnectionHandle.onStateChange()` for tracking connection with remote member in P2P mode. ([#229])
+
+### Changed
+
+- Device audio tracks: ([#225])
+  - `MediaStreamSettings.audio()` renamed to `MediaStreamSettings.deviceAudio()`.
+  - `RoomHandle.muteAudio()`, `RoomHandle.unmuteAudio()`, `RoomHandle.enableAudio()`, `RoomHandle.disableAudio()`, `RoomHandle.enableRemoteAudio()`, and `RoomHandle.disableRemoteAudio()` now receive an argument to specify `MediaSourceKind`.
 
 ### Fixed
 
@@ -28,6 +37,7 @@ All user visible changes to this project will be documented in this file. This p
 - Dependencies:
     - [`medea_flutter_webrtc`] to 0.16.0 version. ([#231])
 
+[#225]: https://github.com/instrumentisto/medea-jason/pull/225
 [#228]: https://github.com/instrumentisto/medea-jason/pull/228
 [#229]: https://github.com/instrumentisto/medea-jason/pull/229
 [#231]: https://github.com/instrumentisto/medea-jason/pull/231